requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
python-dotenv
python-multipart
aiofiles
google-api-python-client
google-auth